Added D Parameter to doubleBootstrap
added logic to doubleBootstrap action to allow a D parameter for size of double bootstraping and to use prior bootstrap samples when they have been created.

Updated walkthroughs and added diagnostic code that looks at data distributions throughout the walkthrough.

Discovered that data is not partitioned so that all of the rows for a bsID or dbsID are on the same host.  Fixed this by adding a partition statement.  The data may be across multiple threads but all rows for single bsid or bsid/dbsid combos will be on the same host now.

updated example 1 to explore all the ways to run the bootstrap and doubleBootstrap actions together or apart.