made jackknife action completely mpp
before this change the key structure for the jackknife was created  on the threads of one worker node.  Now this key is created across all the threads across all workers.

Both version result in the final data being spread evenly across all workers with all data for a single jkID colocated on the same worker.  This version is just more efficient in the creation of the final table.