Stratification Added to resample.bootstrap action (#3)
Incremental Commit names:
* progress

* update for stratification

strata_dev had to be hardcoded in defineActionSet until the binary operater can be figured out.  It was set to 100 so as long as the strata variable has <=100 levels it will work without issue.

* Create casting issue with ceil function.sas

* various update

* updating documentation

* incremental

removed strata from doubleBoostrap to save for later project.  The doubleBootstrap puts a dummy variable for strata in it's single call to bootstrap.

create some starter setup for strata_table in bootstrap walkthrough

* strata setup in bootstrap complete

* Create holdthis

* ready to test bootstrap walkthrough with stratification in external files

* Readme.md Complete For Bootstrap Stratificaiton + New Formatting of Syntax throughout

* Stratified Bootstrap action Complete